whobought-fastapi/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
azure-cosmos==4.3.1
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic>=2.0.0
email-validator>=2.0.0
pyjwt==2.6.0 